refactor: geolocation 공통 hook 생성
