+++ src/hooks/useGeolocationPosition.ts
@@ -1,0 +1,24 @@
+import { useEffect, useState } from 'react';
+
+import { GeolocationCoordinatesType } from '#/types/location';
+
+export function useGeolocationPosition() {
+    const [geolocationCoordinates, setGeolocationCoordinates] =
+        useState<GeolocationCoordinatesType>({ latitude: 0, longitude: 0 });
+
+    useEffect(() => {
+        function handleSuccess(position: GeolocationPosition) {
+            const latitude = position.coords.latitude;
+            const longitude = position.coords.longitude;
+
+            setGeolocationCoordinates({ latitude, longitude });
+        }
+        function handleError(error: GeolocationPositionError) {
+            console.error(error);
+        }
+
+        navigator.geolocation.getCurrentPosition(handleSuccess, handleError);
+    }, []);
+
+    return { geolocationCoordinates };
+}
